feat(game components): Create abstract classes for the game components.
rank.py: create an Enum class for the card ranks (A=1, 2=2, ..., K=13)

suit.py: create an Enum class for the card suits (hearts, clubs, etc.)

card.py: create a class to define what a card is - combination of suit and rank.

deck.py: create a class to define a deck or multiple deck and the respective methods like reset shuffle and deal.

hand.py: define a player or dealer hand and add methods like add card or cards, clear and return the cards in the hand.